Migrate examples/with-refresh-auth to Vite
