+++ examples/with-refresh-auth/README.md
@@ -17,9 +17,9 @@ npm run start
 
 This example contains:
 
-- The `urql` bindings and a React app set up in [`src/App.js`](src/App.js)
-- Some authentication glue code to store the tokens in [`src/auth/Store.js`](src/auth/Store.js)
-- The `Client` and the `authExchange` from `@urql/exchange-auth` set up in [`src/client/index.js`](src/client/index.js)
-- A basic login form in [`src/pages/LoginForm.js`](src/pages/LoginForm.js)
-- And a basic login guard on [`src/pages/Home.js`](src/pages/Home.js)
+- The `urql` bindings and a React app set up in [`src/App.jsx`](src/App.jsx)
+- Some authentication glue code to store the tokens in [`src/authStore.js`](src/authStore.js)
+- The `Client` and the `authExchange` from `@urql/exchange-auth` set up in [`src/client.js`](src/client.js)
+- A basic login form in [`src/pages/LoginForm.jsx`](src/pages/LoginForm.jsx)
+- And a basic login guard on [`src/App.jsx`](src/App.jsx)
   (Note: This isn't using a query in this particular component, since this is just an example)
