+++ examples/with-refresh-auth/package.json
@@ -2,6 +2,9 @@
   "name": "with-refresh-auth",
   "version": "0.0.0",
   "private": true,
+  "scripts": {
+    "start": "vite"
+  },
   "dependencies": {
     "@urql/exchange-auth": "^0.1.2",
     "graphql": "^15.5.0",
@@ -10,24 +13,7 @@
     "urql": "^2.0.2"
   },
   "devDependencies": {
-    "react-scripts": "4.0.3"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "@vitejs/plugin-react-refresh": "^1.3.3",
+    "vite": "^2.2.4"
   }
 }
