--- examples/with-refresh-auth/src/App.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import { Provider } from 'urql';
-
-import client from './client';
-import Home from './pages/Home';
-
-function App() {
-  return (
-    <Provider value={client}>
-      <Home />
-    </Provider>
-  );
-}
-
-export default App;
