+++ examples/with-refresh-auth/src/pages/Profile.jsx
